requirements.txt
pypresence==4.6.1
pyinstaller==6.19.0
Pillow==11.3.0
